Refactor
Refactoring the project for a couple of reasons:

- It was initially a fork that was meant to just fix a single issue with JF, primarily that JF includes empty folders in the library
- I have other scripts I sometimes run against the same library, such as removing unwanted subtitle files (I only care about English ones personally)
- I wanted the overall service to be more flexible in how it operates, so I could scale it up in the future

So the refactor is to update entry.sh to now loop through the scripts.d folder and for every sh in there run them in sequence, each time entry.sh is triggered.  This allows you to fork this and add more scripts as needed, and they'll all run the same way.

I added a new script for my own library to remove non-English subtitle files (which are sometimes created based on how I have tdarr set up).

I updated the existing .ignore file script to be a bit easier to read and manage, just personal preference in how it is done.

I moved the JF API refresh call to entry.sh so it can be triggered after all other library maintenance scripts are run, makes more sense to do it there if it's set up.

Also updated the readme to match what the service does now, and the two included scripts.

**In this specific update I have the subtitle removal script in dry-run mode to test everything, when I'm happy I'll flip it to delete mode by default and merge things in.**